Move NamingScreen in own component file
