+++ front/components/assistant_builder/AssistantBuilder.tsx
@@ -59,17 +59,14 @@ import { useSWRConfig } from "swr";
 
 import { SharingButton } from "@app/components/assistant/Sharing";
 import { TryAssistantModal } from "@app/components/assistant/TryAssistantModal";
-import { AvatarPicker } from "@app/components/assistant_builder/AssistantBuilderAvatarPicker";
 import AssistantBuilderDataSourceModal from "@app/components/assistant_builder/AssistantBuilderDataSourceModal";
 import AssistantBuilderDustAppModal from "@app/components/assistant_builder/AssistantBuilderDustAppModal";
 import AssistantBuilderTablesModal from "@app/components/assistant_builder/AssistantBuilderTablesModal";
 import DataSourceSelectionSection from "@app/components/assistant_builder/DataSourceSelectionSection";
 import DustAppSelectionSection from "@app/components/assistant_builder/DustAppSelectionSection";
 import {
-  DROID_AVATAR_FILES,
-  DROID_AVATARS_BASE_PATH,
-  SPIRIT_AVATAR_FILES,
-  SPIRIT_AVATARS_BASE_PATH,
+  DROID_AVATAR_URLS,
+  SPIRIT_AVATAR_URLS,
   TIME_FRAME_UNIT_TO_LABEL,
 } from "@app/components/assistant_builder/shared";
 import TablesSelectionSection from "@app/components/assistant_builder/TablesSelectionSection";
@@ -97,25 +94,12 @@ import { CONNECTOR_CONFIGURATIONS } from "@app/lib/connector_providers";
 import { isUpgraded } from "@app/lib/plans/plan_codes";
 import { useSlackChannelsLinkedWithAgent, useUser } from "@app/lib/swr";
 import { classNames } from "@app/lib/utils";
+import NamingScreen from "@app/components/assistant_builder/NamingScreen";
 
 type SlackChannel = { slackChannelId: string; slackChannelName: string };
 type SlackChannelLinkedWithAgent = SlackChannel & {
   agentConfigurationId: string;
 };
-
-// Avatar URLs
-const BASE_URL = "https://dust.tt/";
-const buildAvatarUrl = (basePath: string, fileName: string) => {
-  const url = new URL(BASE_URL);
-  url.pathname = `${basePath}${fileName}`;
-  return url.toString();
-};
-const DROID_AVATAR_URLS = DROID_AVATAR_FILES.map((f) =>
-  buildAvatarUrl(DROID_AVATARS_BASE_PATH, f)
-);
-const SPIRIT_AVATAR_URLS = SPIRIT_AVATAR_FILES.map((f) =>
-  buildAvatarUrl(SPIRIT_AVATARS_BASE_PATH, f)
-);
 
 const ACTION_MODE_TO_LABEL: Record<ActionMode, string> = {
   GENERIC: "Reply only",
@@ -1257,114 +1241,6 @@ function AdvancedSettings({
         </div>
       </DropdownMenu.Items>
     </DropdownMenu>
-  );
-}
-
-function NamingScreen({
-  owner,
-  builderState,
-  setBuilderState,
-  setEdited,
-  assistantHandleError,
-}: {
-  owner: WorkspaceType;
-  builderState: AssistantBuilderState;
-  setBuilderState: (
-    statefn: (state: AssistantBuilderState) => AssistantBuilderState
-  ) => void;
-  setEdited: (edited: boolean) => void;
-  assistantHandleError: string | null;
-}) {
-  const [isAvatarModalOpen, setIsAvatarModalOpen] = useState(false);
-
-  return (
-    <>
-      <AvatarPicker
-        owner={owner}
-        isOpen={isAvatarModalOpen}
-        setOpen={setIsAvatarModalOpen}
-        onPick={(avatarUrl) => {
-          setEdited(true);
-          setBuilderState((state) => ({
-            ...state,
-            avatarUrl,
-          }));
-        }}
-        droidAvatarUrls={DROID_AVATAR_URLS}
-        spiritAvatarUrls={SPIRIT_AVATAR_URLS}
-      />
-
-      <div className="flex w-full flex-col gap-4">
-        <div className="text-2xl font-bold text-element-900">Identity</div>
-        <div className="flex flex-row items-start gap-8">
-          <div className="flex flex-col gap-4">
-            <div className="text-lg font-bold text-element-900">Name</div>
-            <div className="flex-grow self-stretch text-sm font-normal text-element-700">
-              Choose a name reflecting the expertise, knowledge access or
-              function of your&nbsp;assistant. Mentioning the&nbsp;assistant in
-              a conversation, like <span className="italic">"@help"</span> will
-              prompt a&nbsp;response from&nbsp;them.
-            </div>
-            <div className="text-sm">
-              <Input
-                placeholder="SalesAssistant, FrenchTranslator, SupportCenter…"
-                value={builderState.handle}
-                onChange={(value) => {
-                  setEdited(true);
-                  setBuilderState((state) => ({
-                    ...state,
-                    handle: value.trim(),
-                  }));
-                }}
-                error={assistantHandleError}
-                name="assistantName"
-                showErrorLabel
-                className="text-sm"
-              />
-            </div>
-            <div className="text-lg font-bold text-element-900">
-              Description
-            </div>
-            <div className="flex-grow self-stretch text-sm font-normal text-element-700">
-              Add a short description that will help Dust and other workspace
-              members understand the&nbsp;assistant’s&nbsp;purpose.
-            </div>
-            <div className="text-sm">
-              <Input
-                placeholder="Answer questions about sales, translate from English to French…"
-                value={builderState.description}
-                onChange={(value) => {
-                  setEdited(true);
-                  setBuilderState((state) => ({
-                    ...state,
-                    description: value,
-                  }));
-                }}
-                error={null} // TODO ?
-                name="assistantDescription"
-                className="text-sm"
-              />
-            </div>
-          </div>
-          <div className="flex flex-col items-center space-y-2">
-            <Avatar
-              size="xl"
-              visual={<img src={builderState.avatarUrl || ""} />}
-            />
-            <Button
-              labelVisible={true}
-              label={"Change"}
-              variant="tertiary"
-              size="xs"
-              icon={PencilSquareIcon}
-              onClick={() => {
-                setIsAvatarModalOpen(true);
-              }}
-            />
-          </div>
-        </div>
-      </div>
-    </>
   );
 }
